requirements.txt
asgiref==3.8.1
asttokens==3.0.0
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.1
daphne==4.1.2
Django==5.1.6
django-channels==0.7.0
djangorestframework==3.15.2
executing==2.2.0
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.68.0
httplib2==0.22.0
hyperlink==21.0.0
icecream==2.1.4
idna==3.10
incremental==24.7.2
msgpack==1.1.0
oauthlib==3.2.2
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-decouple==3.8
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
service-identity==24.2.0
setuptools==75.8.2
six==1.17.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
websocket-client==1.8.0
zope.interface==7.2
